<SU|TS|PY> Added unit tests
