+++ packages/client/src/pages/CarbonHackCards/RegionRecommendationCard/RegionRecommendationCard.tsx
@@ -3,11 +3,14 @@ import { useRegionRecommendationData } from '../../../utils/hooks/RegionRecommen
 import useStyles from '../../EmissionsMetricsPage/CarbonComparisonCard/carbonComparisonStyles'
 import { AWS_REGIONS } from './AWSRegions'
 import DashboardCard from '../../../layout/DashboardCard'
+import Table from '@mui/material/Table'
+import TableBody from '@mui/material/TableBody'
+import TableCell from '@mui/material/TableCell'
+import TableContainer from '@mui/material/TableContainer'
+import TableHead from '@mui/material/TableHead'
+import TableRow from '@mui/material/TableRow'
+import Paper from '@mui/material/Paper'
 import { Typography, CardContent } from '@material-ui/core'
-import { DataGrid, GridColumns } from '@mui/x-data-grid'
-import LoadingMessage from '../../../common/LoadingMessage'
-import ErrorPage from '../../../layout/ErrorPage'
-
 const RegionRecommendationCard: FunctionComponent<any> = ({
   data,
 }): ReactElement => {
@@ -15,6 +18,7 @@ const RegionRecommendationCard: FunctionComponent<any> = ({
 
   const classes = useStyles()
   const footPrintData = data !== undefined ? data : []
+  console.log('Footprint: ', footPrintData)
 
   const findNearestRegions = (
     accountRegion: string,
@@ -27,6 +31,10 @@ const RegionRecommendationCard: FunctionComponent<any> = ({
     } else {
       return ['us-east1', 'us-west1', 'us-east4', 'us-west2']
     }
+  }
+
+  const truncateDecimals = function (number) {
+    return Math.floor(number)
   }
 
   const getNearestRegionFromEnum = (accountRegion: string, REGIONS: any) => {
@@ -42,13 +50,6 @@ const RegionRecommendationCard: FunctionComponent<any> = ({
 
   footPrintData.map((month) => {
     month.serviceEstimates.map((account) => {
-      const date = new Date(
-        new Date(month.timestamp).setMonth(
-          new Date(month.timestamp).getMonth() + 1,
-        ),
-      )
-      const endMonth =
-        date.getMonth() < 9 ? '0' + (date.getMonth() + 1) : date.getMonth() + 1
       accountRegionMap.set(account.accountName, [
         account.region,
         account.cloudProvider,
@@ -56,59 +57,16 @@ const RegionRecommendationCard: FunctionComponent<any> = ({
         account.kilowattHours,
         findNearestRegions(account.region, account.cloudProvider),
         month.timestamp,
-        `${date.getFullYear()}-${endMonth}-${date.getDate()}T00:00:00.000Z`,
+        '2020-02-25',
       ])
     })
   })
 
   const dataToSend = accountRegionMap
+  console.log(dataToSend, ' in emussins card')
 
-  const { result, error, loading } = useRegionRecommendationData(dataToSend)
-
-  const columns: GridColumns = [
-    {
-      field: 'id',
-      headerName: 'Account',
-      width: 265,
-      headerAlign: 'center',
-      align: 'center',
-    },
-    {
-      field: 'region',
-      headerName: 'Region',
-      width: 265,
-      headerAlign: 'center',
-      align: 'center',
-    },
-    {
-      field: 'actualEmissions',
-      headerName: 'Actual Emisssions (Mg)',
-      width: 265,
-      headerAlign: 'center',
-      align: 'center',
-    },
-    {
-      field: 'bestLocation',
-      width: 265,
-      headerName: 'Best Location',
-      headerAlign: 'center',
-      align: 'center',
-    },
-    {
-      field: 'expectedEmissions',
-      headerName: 'Expected Emissions (Mg)',
-      width: 265,
-      headerAlign: 'center',
-      align: 'center',
-    },
-    {
-      field: 'reduction',
-      headerName: 'Reduction in Emissions',
-      width: 265,
-      headerAlign: 'center',
-      align: 'center',
-    },
-  ]
+  const result = useRegionRecommendationData(dataToSend)
+  console.log(result, 'result of emission card')
 
   const bestLocationTable = (result) => {
     // eslint-disable-next-line prefer-const
@@ -116,44 +74,41 @@ const RegionRecommendationCard: FunctionComponent<any> = ({
     for (const [key, value] of result) {
       if (value != undefined) {
         const expectedEmissionsForBestLocation =
-          (value.carbonIntensity * accountRegionMap.get(key)[3]) / 1000000
+          (truncateDecimals(value.carbonIntensity) *
+            truncateDecimals(accountRegionMap.get(key)[3])) /
+          1000000
         const actualEmissions = accountRegionMap.get(key)[2]
-        const percentReduction =
-          100 - (expectedEmissionsForBestLocation * 100) / actualEmissions
-        const reduction =
-          accountRegionMap.get(key)[0] === value.location
-            ? 0
-            : Math.floor(percentReduction)
-        const expectedEmissions =
-          accountRegionMap.get(key)[0] === value.location
-            ? actualEmissions.toFixed(5)
-            : expectedEmissionsForBestLocation.toFixed(5)
-        array.push({
-          id: key,
-          region: accountRegionMap.get(key)[0],
-          actualEmissions: actualEmissions.toFixed(5),
-          bestLocation: value.location,
-          expectedEmissions: expectedEmissions,
-          reduction: `${reduction} %`,
-        })
+        const percentReduction = truncateDecimals(
+          100 - (expectedEmissionsForBestLocation * 100) / actualEmissions,
+        )
+        array.push(
+          <>
+            <TableRow
+              sx={{ '&:last-child td, &:last-child th': { border: 0 } }}
+            >
+              <TableCell align="center">{key}</TableCell>
+              <TableCell component="th" scope="row" align="center">
+                {accountRegionMap.get(key)[0]}
+              </TableCell>
+              <TableCell align="center">{actualEmissions}</TableCell>
+              <TableCell align="center">{value.location}</TableCell>
+              <TableCell align="center">
+                {accountRegionMap.get(key)[0] === value.location
+                  ? actualEmissions
+                  : expectedEmissionsForBestLocation}
+              </TableCell>
+              <TableCell align="center">
+                {accountRegionMap.get(key)[0] === value.location
+                  ? 0
+                  : percentReduction}{' '}
+                %
+              </TableCell>
+            </TableRow>
+          </>,
+        )
       }
     }
     return array
-  }
-
-  const renderTable = () => {
-    return (
-      <div style={{ height: 400, width: '100%' }}>
-        <DataGrid
-          rows={bestLocationTable(result)}
-          columns={columns}
-          pageSize={5}
-          rowsPerPageOptions={[5]}
-          columnBuffer={6}
-          rowBuffer={5}
-        />
-      </div>
-    )
   }
 
   return (
@@ -173,19 +128,21 @@ const RegionRecommendationCard: FunctionComponent<any> = ({
           </Typography>
           <Typography className={classes.posOne}></Typography>
         </CardContent>
-        {error != null ? (
-          <ErrorPage errorMessage={'Error loading cloud data'} />
-        ) : (
-          <>
-            {loading ? (
-              <LoadingMessage
-                message={'Loading cloud data. This may take a while...'}
-              />
-            ) : (
-              renderTable()
-            )}
-          </>
-        )}
+        <TableContainer component={Paper}>
+          <Table sx={{ minWidth: 650 }} aria-label="simple table">
+            <TableHead>
+              <TableRow>
+                <TableCell align="center">Account</TableCell>
+                <TableCell align="center">Region</TableCell>
+                <TableCell align="center">Actual Emissions (Mg)</TableCell>
+                <TableCell align="center">Best Location</TableCell>
+                <TableCell align="center">Expected Emissions (Mg)</TableCell>
+                <TableCell align="center">Reduction in Emissions</TableCell>
+              </TableRow>
+            </TableHead>
+            <TableBody>{bestLocationTable(result)}</TableBody>
+          </Table>
+        </TableContainer>
       </React.Fragment>
     </DashboardCard>
   )
